+++ client/src/App.jsx
@@ -18,70 +18,14 @@ import LoggedOut from "./components/logged-out-page/LoggedOut";
 import SectionList from "./components/main-content/SectionList";
 import Navigation from "./components/outer-components/Navigation";
 import LightBallsOverlay from "./components/visuals/LightBallsOverlay";
+import EnsureUserInDB from "./components/app-components/EnsureUserInDB";
+import UserPage from "./components/app-components/UserPage";
 
 import "./App.css";
 
-const URL = import.meta.env.VITE_API_URL;
 const AUDIENCE = import.meta.env.VITE_API_AUDIENCE;
 const DOMAIN = import.meta.env.VITE_AUTH0_DOMAIN;
 const CLIENTID = import.meta.env.VITE_AUTH0_CLIENT_ID;
-
-const EnsureUserInDB = ({ onReady }) => {
-  const { isAuthenticated, getAccessTokenSilently, user } = useAuth0();
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    if (!isAuthenticated || !user) {
-      setLoading(false);
-      return;
-    }
-    const createUserIfNeeded = async () => {
-      await apiFetch({
-        endpoint: `${URL}/api/users`,
-        method: "POST",
-        body: {
-          email: user.email,
-          name: user.name,
-          username: user.email.split("@")[0],
-        },
-        getAccessTokenSilently,
-      });
-      setLoading(false);
-      onReady && onReady();
-    };
-    createUserIfNeeded();
-  }, [isAuthenticated, getAccessTokenSilently, user, onReady]);
-
-  if (loading) return <div>Loading...</div>;
-  return null;
-};
-
-const UserPage = ({setUserReady, userReady, viewMode, setViewMode }) => {
-  const { user, isLoading, isAuthenticated } = useAuth0();
-  const { username } = useParams();
-  const isOwnPage = user && (username === user.nickname || username === user.name);
-
-  if (isLoading) return <div>Loading...</div>;
-  if (!isAuthenticated || !user) return <LoggedOut />;
-
-  return (
-    <>
-      {isOwnPage && (
-        <EnsureUserInDB onReady={() => setUserReady(true)} />
-      )}
-      {(isOwnPage ? userReady : true) ? (
-        <div className="App">
-          <LightBallsOverlay />
-          <Navigation />
-          <SectionList />
-        </div>
-      ) : (
-        <div>Loading...</div>
-      )}
-    </>
-  );
-}
-
 
 const App = () => {
   const [userReady, setUserReady] = useState(false);
@@ -98,7 +42,6 @@ const App = () => {
     return <LoggedOut />;
   };
 
-  
   return (
     <Auth0Provider
       domain={DOMAIN}
@@ -113,17 +56,17 @@ const App = () => {
     >
       <ViewContext value={{ viewMode, setViewMode }}>
         <Router>
-          
           <Routes>
             <Route path="/" element={<HomeRedirect />} />
             <Route
               path="/:username"
               element={
-                <UserPage 
-                setUserReady={setUserReady} 
-                userReady={userReady}
-                viewMode={viewMode}
-                setViewMode={setViewMode}/>
+                <UserPage
+                  setUserReady={setUserReady}
+                  userReady={userReady}
+                  viewMode={viewMode}
+                  setViewMode={setViewMode}
+                />
               }
             />
             <Route path="/loggedOut" element={<LoggedOut />} />
