+++ server/utils/constants.js
@@ -1,0 +1,17 @@
+const roomActions = Object.freeze({
+  LEAVE: "leave-room",
+  JOIN: "join-room",
+  USERS: "room-users",
+});
+
+const COLORS = [
+  "#ff6b6b",
+  "#ffd93d",
+  "#6bcb77",
+  "#4d96ff",
+  "#a66cff",
+  "#ff6F91",
+  "#ff9671",
+];
+
+module.exports = { roomActions, COLORS };
